eat prompt
