+++ packages/insomnia/package.json
@@ -71,7 +71,6 @@
     "iconv-lite": "^0.6.3",
     "insomnia-plugin-core-themes": "^3.19.0-beta.4",
     "insomnia-plugin-default-headers": "^3.19.0-beta.4",
-    "insomnia-plugin-prompt": "^3.19.0-beta.4",
     "insomnia-plugin-request": "^3.19.0-beta.4",
     "insomnia-plugin-response": "^3.19.0-beta.4",
     "insomnia-testing": "^3.19.0-beta.4",
