--- plugins/insomnia-plugin-prompt/README.md
@@ -1,3 +1,0 @@
-# Insomnia Prompt Template Tag
-
-This is a core Insomnia plugin.
