--- plugins/insomnia-plugin-prompt/package.json
@@ -1,24 +1,0 @@
-{
-  "private": true,
-  "name": "insomnia-plugin-prompt",
-  "version": "3.19.0-beta.4",
-  "author": "Kong <office@konghq.com>",
-  "description": "Insomnia prompt template tag",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/insomnia.git",
-    "directory": "plugins/insomnia-plugin-prompt"
-  },
-  "bugs": {
-    "url": "https://github.com/Kong/insomnia/issues"
-  },
-  "main": "index.js",
-  "insomnia": {
-    "name": "prompt",
-    "unlisted": true
-  },
-  "scripts": {
-    "test": "node --version"
-  }
-}
